requirements/base.txt
Django==1.7.1
#distribute==0.6.34
django-bootstrap-toolkit==2.12.0
python-ldap==2.4.10
pytz==2013b
#South==0.8.1
psycopg2==2.5.1
#django-money==0.4.1
#git+https://github.com/duointeractive/paypal-python.git
dateutils==0.6.6
django-ckeditor==4.4.6
html2text==2014.9.25
beautifulsoup4==4.3.2
Pillow==2.6.0
celery==3.1.17
kombu==3.0.24
billiard==3.3.0.19
amqp==1.4.6
anyjson==0.3.3
django-celery==3.1.16
django-reversion==1.8.5
